@@ -1,46 +1,32 @@
 name: spreadit_crossplatform
-description: "Reddit recreated, but cooler."
+description: "Reddit recreated, but cooler"
 
-publish_to: 'none'
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.0+1
+version: 0.0.1+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ^3.1.1
 
 dependencies:
   flutter:
     sdk: flutter
+  firebase_core: ^2.27.0
+  firebase_auth: ^4.17.8
+  google_sign_in: ^6.2.1
 
-  cupertino_icons: ^1.0.2
-
-  ionicons: ^0.2.2
-
-  flutter_bloc: ^8.1.2
-
-  equatable: ^2.0.5
-
-  get_it: ^7.6.0
-  
-  intl: ^0.19.0
-  
-  retrofit: ^4.1.0
-
-  flutter_hooks: ^0.20.5
-
-  cached_network_image: ^3.2.0
-  
-  dio: ^5.4.1
+  english_words: ^4.0.0
+  provider: ^6.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.1
-  retrofit_generator: ^8.1.0
-  floor_generator: ^1.4.2
-  build_runner: ^2.4.8
-  
+
+  flutter_lints: ^2.0.0
 
 flutter:
-
   uses-material-design: true
+
+  assets:
+  - assets/images/LogoSpreadIt.png
+  - assets/images/GoogleLogo.png
